[Improvement] SwiftUI View 에서 Configuration을 제거
