+++ BezierSwift/Sources/Components/Buttons/Button/BezierButtonStyle.swift
@@ -8,19 +8,21 @@
 import SwiftUI
 
 struct BezierButtonStyle: ButtonStyle {
-  let configuration: BezierButtonConfiguration
+  let cornerRadius: CGFloat
+  let backgroundColor: BezierColor
   
-  init(configuration: BezierButtonConfiguration) {
-    self.configuration = configuration
+  init(cornerRadius: CGFloat, backgroundColor: BezierColor) {
+    self.cornerRadius = cornerRadius
+    self.backgroundColor = backgroundColor
   }
   
   // TODO: Pressed Color 논의가 끝나지 않은 상태라 이번 PR에서 제외합니다. by Tom 2024.08.02
   func makeBody(configuration: Configuration) -> some View {
     configuration.label
       .background(
-        RoundedRectangle(cornerRadius: self.configuration.cornerRadius, style: .circular)
+        RoundedRectangle(cornerRadius: self.cornerRadius, style: .circular)
           .foregroundColor(
-            configuration.isPressed ? self.configuration.backgroundColor.color: self.configuration.backgroundColor.color
+            configuration.isPressed ? self.backgroundColor.color: self.backgroundColor.color
           )
       )
   }
